MDL-81055 lib: Allow dir attr through purify_html
The 'dir' attribute can now be included with multilang strings.
The changes here are in two areas. Firstly, ensure purify_html does
not strip or convert the new dir attribute. This is the same solution
that was used for allowing the 'lang' attribute through untouched.

Secondly, the multilang filter can match for 'dir' too, allowing the
retrieval of the desired language value if this attribute is present.